End of lessons
In some schools, classes start at 9:00. The duration of the lesson is 45 minutes, after the 1st, 3rd, 5th, etc. lessons change 5 minutes, and after the 2nd, 4th, 6th, etc. — 15 minutes.

The lesson number is given (a number from 1 to 10). Determine when the specified lesson ends.

Print two integers: the end time of the lesson in hours and minutes.